Исправлены замечания: разбиты селекторы, добавлено действие resetCurrentFilmData
