+++ src/components/add-review-screen/add-review-screen.jsx
@@ -5,11 +5,10 @@ import ReviewForm from "../review-form/review-form";
 import FilmHeader from "../film-header/film-header";
 import PosterSmall from "../poster/poster-small";
 import {useDispatch, useSelector} from "react-redux";
-import {getCurrentFilmFromFilmsList} from "../../store/data/selectors";
 import {changeCurrentFilmID} from "../../store/action-creator";
 import NotFoundScreen from "../not-found-screen/not-found-screen";
 import Spinner from "../spinner/spinner";
-import {needShowSpinnerInsteadCurrentFilm} from "../../store/data/selectors-with-loading-status";
+import {getCurrentFilmFromFilmsList, needShowSpinnerInsteadCurrentFilm} from "../../store/data/selectors/current-film";
 
 const AddReviewScreen = ({currentFilmID}) => {
   const currentFilm = useSelector(getCurrentFilmFromFilmsList);
