+++ src/components/catalog-main/catalog-main.jsx
@@ -2,16 +2,16 @@ import React, {useEffect, useState} from "react";
 import Films from "../films/films";
 import Spinner from "../spinner/spinner";
 import GenresList from "../genres-list/genres-list";
-import {getFilmsByActiveGenre} from "../../store/data/selectors";
 import {useSelector} from "react-redux";
-import {needShowSpinnerInsteadFilms} from "../../store/data/selectors-with-loading-status";
 import ShowMoreButton from "../show-more-button/show-more-button";
+import {getFilmsByActiveGenre} from "../../store/data/selectors/films";
+import {needShowSpinnerInsteadFilms} from "../../store/data/selectors/selectors";
 
 const INITIAL_SHOWN_FILMS_COUNTER = 8;
 
 const CatalogMain = () => {
-  const isSpinnerShown = useSelector(needShowSpinnerInsteadFilms);
   const films = useSelector(getFilmsByActiveGenre);
+  const isSpinnerShown = useSelector(needShowSpinnerInsteadFilms);
 
   const [isShowMoreButtonShown, setIsShowMoreButtonShown] = useState(true);
   const [shownFilmsCounter, setShownFilmsCounter] = useState(INITIAL_SHOWN_FILMS_COUNTER);
