+++ src/components/film-screen/film-screen.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import Tabs from "../tabs/tabs";
 import CatalogSimilar from "../catalog-similar/catalog-similar";
 import {getAuthorizationStatus} from "../../store/user/selectors";
-import {getCurrentFilm, getCurrentFilmReviews} from "../../store/data/selectors";
 import {useDispatch, useSelector} from "react-redux";
 import Poster from "../poster/poster";
 import FilmInfo from "../film-info/film-info";
@@ -13,7 +12,8 @@ import {changeCurrentFilmID} from "../../store/action-creator";
 import NotFoundScreen from "../not-found-screen/not-found-screen";
 import Spinner from "../spinner/spinner";
 import {fetchFilm, fetchReviews} from "../../store/api-actions";
-import {needShowSpinnerInsteadCurrentFilm} from "../../store/data/selectors-with-loading-status";
+import {getCurrentFilmReviews, needShowSpinnerInsteadCurrentFilm} from "../../store/data/selectors/current-film";
+import {getCurrentFilm} from "../../store/data/selectors/selectors";
 
 const FilmScreen = ({currentFilmID}) => {
   const currentFilm = useSelector(getCurrentFilm);
